Add new api for get variations of a product.

+++ src/main/java/com/prico/dto/comparison/ProductStoreDto.java
@@ -19,5 +19,5 @@ public class ProductStoreDto {
 
     private String imageUrl;
 
-    private Long price;
+    private Float price;
 }
